Implemented Peter's suggestions
