+++ s2spy/rgdr/__init__.py
@@ -1,4 +1,3 @@
 """Response Guided Dimensionality Reduction."""
-from . import _map_analysis
-from . import _map_regions
+from . import _rgdr
 from ._rgdr import RGDR
